Bugfix in adelay_update and IPv4::transmit. Continued work on TCP retransmission cache.

+++ src/server/adelay.c
@@ -1,10 +1,10 @@
 #include <DPA/UCS/adelay.h>
 
-static unsigned last;
+static unsigned long last;
 static adelay_t time;
 
 void adelay_update_time( unsigned long current, unsigned long max_mask, unsigned long ticks_per_sec ){
-  unsigned passed = (unsigned long)( ( current - last ) & max_mask ) * AD_SEC / ticks_per_sec;
+  unsigned long passed = (unsigned long)( ( current - last ) & max_mask ) * AD_SEC / ticks_per_sec;
   if( !passed )
     return;
   last += (unsigned long)passed * ticks_per_sec / AD_SEC;
